ootb-*: add build.4 to address issue w/ outputs
before this build, cartographer would pass forward outputs that are
completely empty if, e.g., it found the field that an `imagePath`
describes (like, `.status.foo`) but that field would be empty (like,
`.status.foo: ""`), a situation that would occur for `ImageScan`.

in this build, we update the iamge scan template so that we use a field
where cartographer wouldn't see an empty value like that to promote
forward.

fixed in https://gitlab.eng.vmware.com/tap-supply-chains/catalog/-/merge_requests/14

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>