ootb-templates: exclude tkn tasks in run profile
regardless of which profile gets configured for installation,
`ootb-templates` is installed as it brings all the necessary bits for
the supply chains to work, e.g., the shared templates for instantiating
git repositories, image repositories, etc, as well as tekton tasks for
pushing to git repositories or image registries.

the latter (tekton tasks) are not desired in the case of a run profile
though - a run cluster will never have tekton installed.

given that `ootb_templates` support the configuration of
`excluded_tempaltes`, an array that lets you tell which templates should
_not_ be submitted to the cluster at installation time, here we do a
condition on the profile to automatically fill the array with tekton
tasks so that in the case of a Run profile being installed, those are
not submitted.

e.g.:

	profile: run

		ytt --data-value profile=run \
			-f ./_profiles.star \
			-f ./values.yml \
			-f ./ootb-templates.yaml

		...
		stringData:
		  values.yml: |
		    excluded_templates:
		    - git-writer
		    - image-writer

	profile: full	(could be any other)

		...
		stringData:
		  values.yml: |
		    {}

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>