Add ca-overlay to convention-controller 0.6.3 package
The package definition was missing a reference to an overlay that
applied the ca-cert-data from the install values into the resources
being installed. Without this overlay it's as if no ca certs were
defined.

The additional overlay was part of our packages own CI tests, the entry
was omitted because I duplicated a previous package definition and
updated the bundle reference rather than generate a clean new package.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>