cartographer: add v0.4.0-build.3
in this release of cartographer, for the first time we make use of the
new packaging from `vmware-tanzu/package-for-cartographer` where both
`cartographer` _and_ `cartographer-conventions` (umbrella project) are
packaged together.

as a result, the Package object for this release include two extra
properties:

- `ca_cert_data`, for passing to cartographer-conventions pem-encoded
  ca-certificates to be trusted by the conventions controllers
- `excluded_components`, for letting one to avoid installing either
  `cartographer` or `conventions` (as they're both packaged together).

note that `cartographer-conventions` makes use of a new group version
(conventions.carto.run, as opposed to
conventions.apps.tanzu.vmware.com), but given that
`convention-controller` (from the `convention-controller` package)
provides a shim that translates from the old crd to the new ones, the
transition should be smooth.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>