Reformat for easier automation
We are using ytt to automate the changes we make to the tap-packages
repo for new versions of API portal.

When we do this, the repos/<tap-version>.yaml file loses a bunch of
whitespace and this makes it hard to confirm that we have only added our
newest version to the api portal package.

This change will allow the diff to be much smaller when we add our one
line and will speed up the integration of API portal into TAP.